requirements.txt
streamlit
pandas
supabase
requests
xlsxwriter
plotly
google-generativeai>=0.8.3
streamlit-aggrid  <-- 新增这一行